+++ main.py
@@ -5,14 +5,14 @@ from flask import Flask, request
 from flask_restful import Api, Resource
 from loguru import logger
 
-from Types import AddInfo, RequestPayload
 from _logging import CONSOLE_LOGGING_CONFIG, FILE_LOGGING_CONFIG
-from exceptions import BackendUnreachableError
 from feecc_spoke import Alerts, Views
 from feecc_spoke.Display import Display
 from feecc_spoke.Employee import Employee
+from feecc_spoke.Exceptions import BackendUnreachableError
 from feecc_spoke.Spoke import Spoke
 from feecc_spoke.State import AuthorizedIdling, AwaitLogin, ProductionStageOngoing
+from feecc_spoke.Types import AddInfo, RequestPayload
 
 # apply logging configuration
 logger.configure(handlers=[CONSOLE_LOGGING_CONFIG, FILE_LOGGING_CONFIG])
